hexdump-fix-for-non-aligned-buffers-v6
 - Removed unnecessary cast of u64 to unsigned long long, since
     currently they are equivalent (per Geer Uytterhoeven's indication)

Cc: Geert Uytterhoeven <geert@linux-m68k.org>

+++ lib/hexdump.c
@@ -140,8 +140,7 @@ int hex_dump_to_buffer(const void *buf, size_t len, int rowsize, int groupsize,
 		for (j = 0; j < ngroups; j++) {
 			ret = snprintf(linebuf + lx, linebuflen - lx,
 				       "%s%16.16llx", j ? " " : "",
-				       (unsigned long long)
-					       get_unaligned(ptr8 + j));
+				       get_unaligned(ptr8 + j));
 			if (ret >= linebuflen - lx)
 				goto overflow1;
 			lx += ret;
